pipeline/requirements.txt
musicbrainzngs==0.7.1
pycountry==24.6.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
spotipy==2.24.0
